pyproject.toml
[tool.poetry]
name = "reporover"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1"
pandas = "^1.2.4"
PyDriller = "^1.15.5"
PyGithub = "^1.55"
giturlparse = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
